feat(network-manager): add systemd generator if available
A systemd generator was added in [1] to fix an upstream issue [2]. It also
brings the removal of the "[Install]" section from the upstream initrd services;
it is used to mask the initrd services on the host, and to mask the host
services and provide installation links of the initrd services in the initrd.

Additionally, increase the minimum NetworkManager version set to remove custom
systemd services from dracut.

[1] https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/commit/636fb5ef24640856515584977174fa44a986e374
[2] https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/issues/1814

(cherry picked from commit b6329a088e2c49ee749361a1401e98380478da26)